Makefile: notmuch.vim and notmuch.txt live in subdirs
Both files were moved out of the top-level directory into subdirectories
in dfb5da86df23, "Reorganize files" which was introduced in response to
issue #5, "compatibility with Vundle".

Makefile was not updated causing the following error:

  $ make install
  install -v -D -m644 /src/notmuch-vim/plugin/ /dest/.vim/plugin/
  install: omitting directory ‘/src/notmuch-vim/plugin/’
  make: *** [install] Error 1

Additionally the method used to populate the plugin and doc directories
was updated to iterate over all files in those directories. This
future-proofs those directories in the event that new files are someday
added to them.

Signed-off-by: Michael Turquette <mturquette@deferred.io>